requirements.txt
Django>=4.2.0
django-ratelimit>=4.1.0
celery>=5.3.0
redis>=4.6.0
django-ipware>=5.0.0
scikit-learn>=1.3.0
django-celery-beat>=2.6.0
requests>=2.31.0
